+++ vite-plugin-ssr/node/runtime/html/injectAssets.ts
@@ -5,7 +5,7 @@ export type { PageContextPromise }
 
 import { assert, isCallable, isPromise } from '../utils'
 import type { PageAsset } from '../renderPage/getPageAssets'
-import { assertPageContextProvidedByUser } from '../../shared/assertPageContextProvidedByUser'
+import { assertPageContextProvidedByUser } from '../../../shared/assertPageContextProvidedByUser'
 import { injectHtmlTags, createHtmlHeadIfMissing } from './injectAssets/injectHtmlTags'
 import type { InjectToStream } from 'react-streaming/server'
 import type { HtmlPart } from './renderHtml'
