+++ vite-plugin-ssr/node/runtime/page-files/setup.ts
@@ -1,4 +1,4 @@
-import { setPageFilesAsync } from '../../shared/getPageFiles'
+import { setPageFilesAsync } from '../../../shared/getPageFiles'
 import { assert, debugGlob } from '../utils'
 import { getGlobalContext } from '../globalContext'
 import { virtualModuleIdPageFilesServer } from '../plugin/plugins/generateImportGlobs/virtualModuleIdPageFiles'
