+++ vite-plugin-ssr/node/runtime/renderPage.ts
@@ -7,9 +7,9 @@ import {
   renderPageContext,
   RenderResult
 } from './renderPage/renderPageContext'
-import { route } from '../shared/route'
+import { route } from '../../shared/route'
 import { assert, hasProp, objectAssign, isParsable, parseUrl } from './utils'
-import { addComputedUrlProps } from '../shared/addComputedUrlProps'
+import { addComputedUrlProps } from '../../shared/addComputedUrlProps'
 import { isRenderErrorPageException } from './renderPage/RenderErrorPage'
 import { initGlobalContext } from './globalContext'
 import { handlePageContextRequestUrl } from './renderPage/handlePageContextRequestUrl'
