+++ vite-plugin-ssr/node/runtime/renderPage/getPageAssets.ts
@@ -19,7 +19,7 @@ import path from 'path'
 import { inferMediaType, type MediaType, type ViteManifest } from '../helpers'
 import { getManifestEntry } from './getPageAssets/getManifestEntry'
 import type { ViteDevServer } from 'vite'
-import type { ClientDependency } from '../../shared/getPageFiles/analyzePageClientSide/ClientDependency'
+import type { ClientDependency } from '../../../shared/getPageFiles/analyzePageClientSide/ClientDependency'
 import { sortPageAssetsForEarlyHintsHeader } from './getPageAssets/sortPageAssetsForEarlyHintsHeader'
 import type { ConfigVpsResolved } from '../plugin/plugins/config/ConfigVps'
 import { getGlobalContext } from '../globalContext'
