+++ vite-plugin-ssr/node/runtime/renderPage/getPageAssets/retrieveAssetsProd.ts
@@ -4,7 +4,7 @@ import { assert, isNpmPackageModule } from '../../utils'
 import { ViteManifest } from '../../helpers'
 import { getManifestEntry } from './getManifestEntry'
 import { extractAssetsAddQuery } from '../../plugin/plugins/extractAssetsPlugin/extractAssetsAddQuery'
-import type { ClientDependency } from '../../../shared/getPageFiles/analyzePageClientSide/ClientDependency'
+import type { ClientDependency } from '../../../../shared/getPageFiles/analyzePageClientSide/ClientDependency'
 
 function retrieveAssetsProd(
   clientDependencies: ClientDependency[],
