+++ vite-plugin-ssr/node/runtime/renderPage/loadPageFilesServer.ts
@@ -2,11 +2,11 @@ export { loadPageFilesServer }
 export type { PageFiles }
 export type { PageContext_loadPageFilesServer }
 
-import { type PageFile, getExportUnion } from '../../shared/getPageFiles'
-import { analyzePageClientSide, analyzePageClientSideInit } from '../../shared/getPageFiles/analyzePageClientSide'
+import { type PageFile, getExportUnion } from '../../../shared/getPageFiles'
+import { analyzePageClientSide, analyzePageClientSideInit } from '../../../shared/getPageFiles/analyzePageClientSide'
 import { assertWarning, objectAssign, PromiseType } from '../utils'
 import { getPageAssets, PageContextGetPageAssets, type PageAsset } from './getPageAssets'
-import { loadPageFilesServerSide } from '../../shared/getPageFiles/analyzePageServerSide/loadPageFilesServerSide'
+import { loadPageFilesServerSide } from '../../../shared/getPageFiles/analyzePageServerSide/loadPageFilesServerSide'
 import { debugPageFiles, type PageContextDebug } from './debugPageFiles'
 import type { MediaType } from '../helpers'
