+++ vite-plugin-ssr/node/runtime/renderPage/preparePageContextForRelease.ts
@@ -1,12 +1,12 @@
 export { preparePageContextForRelease }
 export type { PageContextPublic }
 
-import { isErrorPageId } from '../../shared/route'
-import type { ExportsAll } from '../../shared/getPageFiles'
+import { isErrorPageId } from '../../../shared/route'
+import type { ExportsAll } from '../../../shared/getPageFiles'
 import { assert, hasProp, isPlainObject, isObject } from '../utils'
-import { sortPageContext } from '../../shared/sortPageContext'
+import { sortPageContext } from '../../../shared/sortPageContext'
 import { addIs404ToPageProps } from '../helpers'
-import { assertURLs, PageContextUrls } from '../../shared/addComputedUrlProps'
+import { assertURLs, PageContextUrls } from '../../../shared/addComputedUrlProps'
 
 type PageContextPublic = {
   urlOriginal: string
